+++ Source/SourceKittenFramework/Request.swift
@@ -91,6 +91,10 @@ private func fromSourceKit(sourcekitObject: sourcekitd_variant_t) -> SourceKitRe
 /// dispatch_once_t token used to only initialize SourceKit once per session.
 private var sourceKitInitializationToken: dispatch_once_t = 0
 
+/// dispatch_semaphore_t used when waiting for sourcekitd to be restored.
+private var sourceKitWaitingRestoredSemaphore = dispatch_semaphore_create(0)
+private let sourceKitWaitingRestoredTimeout = Int64(10 * NSEC_PER_SEC)
+
 /// SourceKit UID to String map.
 private var uidStringMap = [sourcekitd_uid_t: String]()
 
@@ -241,7 +245,7 @@ public enum Request {
             return nil
         }
         sourcekitd_request_dictionary_set_int64(request, sourcekitd_uid_get_from_cstr(SwiftDocKey.Offset.rawValue), offset)
-        return Request.CustomRequest(request).send()
+        return try? Request.CustomRequest(request).failableSend()
     }
 
     /**
@@ -255,6 +259,72 @@ public enum Request {
         }
         let response = sourcekitd_send_request_sync(sourcekitObject)
         defer { sourcekitd_response_dispose(response) }
+        return fromSourceKit(sourcekitd_response_get_value(response)) as! [String: SourceKitRepresentable]
+    }
+    
+    /// A enum representation of SOURCEKITD_ERROR_*
+    public enum Error: ErrorType, CustomStringConvertible {
+        case ConnectionInterrupted(String?)
+        case Invalid(String?)
+        case Failed(String?)
+        case Cancelled(String?)
+        case Unknown(String?)
+        
+        /// A textual representation of `self`.
+        public var description: String {
+            return getDescription() ?? "no description"
+        }
+        
+        private func getDescription() -> String? {
+            switch self {
+            case .ConnectionInterrupted(let string): return string
+            case .Invalid(let string): return string
+            case .Failed(let string): return string
+            case .Cancelled(let string): return string
+            case .Unknown(let string): return string
+            }
+        }
+        
+        private init(response: sourcekitd_response_t) {
+            let description = String(UTF8String: sourcekitd_response_error_get_description(response))
+            switch sourcekitd_response_error_get_kind(response) {
+            case SOURCEKITD_ERROR_CONNECTION_INTERRUPTED: self = .ConnectionInterrupted(description)
+            case SOURCEKITD_ERROR_REQUEST_INVALID: self = .Invalid(description)
+            case SOURCEKITD_ERROR_REQUEST_FAILED: self = .Failed(description)
+            case SOURCEKITD_ERROR_REQUEST_CANCELLED: self = .Cancelled(description)
+            default: self = .Unknown(description)
+            }
+        }
+    }
+    
+    /**
+    Sends the request to SourceKit and return the response as an [String: SourceKitRepresentable].
+     
+    - returns: SourceKit output as a dictionary.
+    - throws: Request.Error on fail ()
+    */
+    public func failableSend() throws -> [String: SourceKitRepresentable] {
+        dispatch_once(&sourceKitInitializationToken) {
+            sourcekitd_initialize()
+            sourcekitd_set_notification_handler() { response in
+                if !sourcekitd_response_is_error(response) {
+                    fflush(stdout)
+                    fputs("sourcekitten: connection to SourceKitService restored!\n", stderr)
+                    dispatch_semaphore_signal(sourceKitWaitingRestoredSemaphore)
+                }
+                sourcekitd_response_dispose(response)
+            }
+        }
+        let response = sourcekitd_send_request_sync(sourcekitObject)
+        defer { sourcekitd_response_dispose(response) }
+        if sourcekitd_response_is_error(response) {
+            let error = Request.Error(response: response)
+            if case .ConnectionInterrupted = error {
+                dispatch_semaphore_wait(sourceKitWaitingRestoredSemaphore,
+                    dispatch_time(DISPATCH_TIME_NOW, sourceKitWaitingRestoredTimeout))
+            }
+            throw error
+        }
         return fromSourceKit(sourcekitd_response_get_value(response)) as! [String: SourceKitRepresentable]
     }
 }