+++ Source/sourcekitten/main.swift
@@ -7,18 +7,25 @@
 //
 
 import Darwin
+import Foundation
 import Commandant
 
-let registry = CommandRegistry<SourceKittenError>()
-registry.register(CompleteCommand())
-registry.register(DocCommand())
-registry.register(SyntaxCommand())
-registry.register(StructureCommand())
-registry.register(VersionCommand())
+// `sourcekitd_set_notification_handler()` set the handler to be executed on main thread queue.
+// So, we vacate main thread to `dispatch_main()`.
+dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0)) {
+    let registry = CommandRegistry<SourceKittenError>()
+    registry.register(CompleteCommand())
+    registry.register(DocCommand())
+    registry.register(SyntaxCommand())
+    registry.register(StructureCommand())
+    registry.register(VersionCommand())
 
-let helpCommand = HelpCommand(registry: registry)
-registry.register(helpCommand)
+    let helpCommand = HelpCommand(registry: registry)
+    registry.register(helpCommand)
 
-registry.main(defaultVerb: "help") { error in
-    fputs("\(error)\n", stderr)
+    registry.main(defaultVerb: "help") { error in
+        fputs("\(error)\n", stderr)
+    }
 }
+
+dispatch_main()