+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ##### Breaking
 
-* None.
+* Change `SwiftDocs.init(file:arguments:)` to
+  `SwiftDocs.init?(file:arguments:)`  
+  [Norio Nomura](https://github.com/norio-nomura)
 
 ##### Enhancements
 
-* None.
+* Add `Request.failableSend()` that can handle SourceKitService crashes.
+  `sourcekitten doc` does not stop when SourceKitService crashes.  
+  [Norio Nomura](https://github.com/norio-nomura)
 
 ##### Bug Fixes